Added first Code First Migration.
Updated FriendOrganizer.DataAccess\App.config to have a connection string of FriendOrganizerDb which is the string we used when calling the DbContext base constructor in FriendOrganizerDbContext. In the nuget package manager console we used "Enable-Migrations" for DataAccess. We added a seed method to the generated Configuration.cs file in Migrations to add or update the database with hard coded values. Then we called "Add-Migration InitialDatabase" with the latter being the name of the created database. This created a new database cs file in the Migrations folder.