Create SQL Server and seed with friend data.
I have had to deviate from the tutorial and use data source (localdb)\MSSQLLocalDB. I add this to the DataAccess projects App.config as the default entity framework code first project with migrations enabled, and to the UI project as it is the startup project. I then run on the package manager console "Update-Database". I then create a Data Connection in the Server Explorer with server as (localdb)\MSSQLLocalDB and database as FriendOrganizer. I can then view the database data in SeverExplorer/Data Connections/Tables.