Moved text boxes for friends to FriendDetailView and ViewModel.
Created FriendDetailViewModel class which takes a data service as a parameter. It has a load async method to call the data service and a Friend private property. We extract the interface and register it with the bootstrapper. We add an instance of the interface to the MainViewModel. We move the XAML code for the friend details to a FriendDetailView user interface file and call it from the MainWindow, giving FriendDetailViewModel as the DataContext.