Using events to communicate between Navigation and FriendDetail.
Added SelectedFriend property to NavigationViewModel and bound this to the NavigationView. The SelectedFriend publishes to the OpenFriendDetailViewEvent. FriendDetailViewModel subscribes to this event, calling the OnOpenFriendDetailView which calls the LoadAsync method which sets the Friend property, which fills the textboxes with the relevant details. Altered FriendDataService to return Friend rather than list when passed the friendId as a parameter.