Created warning message box when leaving unsaved changes.
Created View/Services folder with a message dialog service, to keep the message popups out of the view model for testing. I have created a MessageDialogService class and extracted an interface for displaying a popup message. I have implemented the class in the main view model to prevent immediate navigation away from changes and registered the class and interface in the bootstrapper. I have had to add a HasChanges property to the IFriendDetailViewModel interface.