Added Create Friend option.
Added grid rows to MainWindow for menu bar.  Added a command CreateNewFriendCommand to MenuItem. Created this DelegateCommand in MainViewModel and set a new method OnCreateNewFriendExecute as its param. This calls OnOpenNewFriend with value null for friendId. friendId and all the dependents are set to nullable int? types. FriendDetailView needs a new CreateNewFriend method in the case LoadAsync is called with null friendId. This requires FriendRepository to have a new Add method to add the friend to its context. AfterFriendSaved in NavigationViewModel must be altered to allow creation of a new NavigationItemViewModel for the new friend.